清理 GUI 无引用兼容入口
